One example working with liveness for generators
This patch uses liveness to do generator interior analysis for an
example program.

The following program now passes the type checker, while it didn't
before:

    fn main() {
        let g = move || {
            match status(&Client) {
                _status => yield,
            }
        };
        assert_send(g);
    }

In contrast, the following program does not pass because &Client is
used after the yield.

    fn main() {
        let g = move || {
            let client = &Client;
            match status(client) {
                _status => yield,
            }
            status(client)
        };
        assert_send(g);
    }

Note that this version is not sound, but it hints that the liveness
approach will work.